Refactor WalletCoreProvider imports and update yarn.lock
- Updated imports for WalletCoreProvider from '@core/chain-ui' to '@core/ui' across multiple components and pages.
- Added new package '@core/ui' in yarn.lock with dependencies for WalletCore.
- Removed deprecated entries from yarn.lock related to '@core/chain-ui'.
