+++ clients/desktop/src/vault/deposit/DepositConfirmButton/index.tsx
@@ -2,9 +2,9 @@ import { create } from '@bufbuild/protobuf'
 import { toChainAmount } from '@core/chain/amount/toChainAmount'
 import { Chain } from '@core/chain/Chain'
 import { coinKeyFromString } from '@core/chain/coin/Coin'
-import { useAssertWalletCore } from '@core/chain-ui/providers/WalletCoreProvider'
 import { toCommCoin } from '@core/mpc/types/utils/commCoin'
 import { KeysignPayloadSchema } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
+import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
 import { isOneOf } from '@lib/utils/array/isOneOf'
 import { shouldBePresent } from '@lib/utils/assert/shouldBePresent'
 import { useTranslation } from 'react-i18next'
