+++ clients/desktop/src/vault/keysign/shared/mutations/useKeysignMutation.ts
@@ -7,10 +7,10 @@ import { executeTx } from '@core/chain/tx/execute'
 import { getPreSigningHashes } from '@core/chain/tx/preSigningHashes'
 import { generateSignature } from '@core/chain/tx/signature/generateSignature'
 import { hexEncode } from '@core/chain/utils/walletCore/hexEncode'
-import { useAssertWalletCore } from '@core/chain-ui/providers/WalletCoreProvider'
 import { KeysignMessagePayload } from '@core/mpc/keysign/keysignPayload/KeysignMessagePayload'
 import { MPCKeysign } from '@core/mpc/mpcKeysign'
 import { MpcLib } from '@core/mpc/mpcLib'
+import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
 import { getLastItem } from '@lib/utils/array/getLastItem'
 import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
 import { chainPromises } from '@lib/utils/promise/chainPromises'
