+++ clients/desktop/src/vault/keysign/start/fast/FastKeysignServerStep.tsx
@@ -3,7 +3,7 @@ import { getCoinType } from '@core/chain/coin/coinType'
 import { signatureAlgorithms } from '@core/chain/signing/SignatureAlgorithm'
 import { getPreSigningHashes } from '@core/chain/tx/preSigningHashes'
 import { hexEncode } from '@core/chain/utils/walletCore/hexEncode'
-import { useAssertWalletCore } from '@core/chain-ui/providers/WalletCoreProvider'
+import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
 import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
 import { assertField } from '@lib/utils/record/assertField'
 import { useMutation } from '@tanstack/react-query'
