+++ clients/desktop/src/vault/mutations/useUpdateAddressBookItemMutation.ts
@@ -1,5 +1,5 @@
 import { Chain } from '@core/chain/Chain'
-import { useAssertWalletCore } from '@core/chain-ui/providers/WalletCoreProvider'
+import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 
 import { UpdateAddressBookItem } from '../../../wailsjs/go/storage/Store'
