+++ clients/desktop/src/vault/swap/queries/useSwapKeysignPayloadQuery.ts
@@ -1,8 +1,8 @@
 import { create } from '@bufbuild/protobuf'
 import { toChainAmount } from '@core/chain/amount/toChainAmount'
-import { useAssertWalletCore } from '@core/chain-ui/providers/WalletCoreProvider'
 import { toCommCoin } from '@core/mpc/types/utils/commCoin'
 import { KeysignPayloadSchema } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
+import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
 
 import { processKeysignPayload } from '../../../chain/keysign/processKeysignPayload'
 import { getSwapKeysignPayloadFields } from '../../../chain/swap/keysign/getSwapKeysignPayloadFields'
