+++ core/ui/chain/providers/WalletCoreProvider.tsx
