+++ core/ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@core/chain-ui",
+  "name": "@core/ui",
   "version": "0.0.1",
   "private": true,
   "scripts": {
